--- a/Gemfile
+++ b/Gemfile
@@ -32,7 +32,8 @@ gem "kaminari"
 gem "groupdate"
 # Pour la validation avancée des mots de passe
 gem "zxcvbn-ruby"
-
+# Pour parraléliser
+gem 'concurrent-ruby'
 
 # Use Active Model has_secure_password [https://guides.rubyonrails.org/active_model_basics.html#securepassword]
 # gem "bcrypt", "~> 3.1.7"
@@ -70,9 +71,7 @@ group :development, :test do
 end
 
 group :development do
-  # Use console on exceptions pages [https://github.com/rails/web-console]
   gem "web-console"
-  # Pour la gestion des emails en développement (optionnel)
   gem "letter_opener"
 end