… children and schooling tabs

Performance improvements:
- Parallelize API calls in load_children_data_with_cache using concurrent-ruby
  * Children tab: 13s -> 6.3s (52% faster)
- Parallelize API calls in load_schooling_data_with_cache using concurrent-ruby
  * Schooling tab: 10.4s -> 6.8s (35% faster)
- API calls to EPCI, department, region, and France data now execute concurrently instead of sequentially

Performance monitoring system:
- Add real-time performance tracking for all dashboard tabs in AsyncSectionLoader
- Display colored badges (green/orange/red) based on load times
- Store last 10 metrics per section for analysis
- Add performance summary methods to EpciDashboardManager
- Console commands: showPerformanceSummary(), getPerformanceStats(), togglePerformanceMonitoring()
- Configurable thresholds: fast (<2s), medium (2-5s), slow (>5s)

Dependencies:
- Add concurrent-ruby gem for parallel promise execution

This significantly improves user experience for the slowest dashboard sections and provides ongoing visibility into performance metrics.